Refactor : DTO구조 변경
