+++ src/routes/search/search.res.dto.ts
@@ -1,7 +1,3 @@
-/**
- * @pattern ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
- */
-
 import { UUID } from '../../@types/common.js';
 
 export interface SearchResDTO {
