+++ src/routes/wish/dto/wish.req.dto.ts
@@ -1,0 +1,13 @@
+import { UUID } from '../../../@types/common.js';
+
+export type WishType = 'PROPOSAL' | 'ITEM' | 'REQUEST';
+
+export interface WishReqDTO {
+  /**
+   * 위시 타입
+   * @example PROPOSAL
+   */
+  type: WishType;
+
+  itemId: UUID;
+}
