+++ src/routes/wish/dto/wish.res.dto.ts
@@ -1,37 +1,14 @@
-/**
- * @pattern ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
- */
-
-import { UUID } from '../../@types/common.js';
+import { UUID } from '../../../@types/common.js';
+import { WishType } from './wish.req.dto.js';
 
 export interface WishResDTO {
-  /**
-   * @format uuid
-   */
   wishId: UUID;
   createdAt: Date;
 }
 
 export interface DeleteWishResDTO {
-  /**
-   * @format uuid
-   */
   wishId: UUID;
   deletedAt: Date;
-}
-
-export type WishType = 'PROPOSAL' | 'ITEM' | 'REQUEST';
-
-export interface WishReqDTO {
-  /**
-   * 위시 타입
-   * @example PROPOSAL
-   */
-  type: WishType;
-  /**
-   * @format uuid
-   */
-  itemId: UUID;
 }
 
 export interface WishDetailDTO {
